% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-package.R
\docType{data}
\name{ball_bearing}
\alias{ball_bearing}
\alias{ball.bearing}
\alias{ballBearing}
\title{Lifespan of ball bearings}
\format{A tibble with 75 observations on the following variable.
\describe{
\item{life_span}{Lifespan of ball bearings (in hours).}
}}
\source{
Simulated data.
}
\usage{
ball_bearing
}
\description{
A simulated data set on lifespan of ball bearings.
}
\examples{

histPlot(ball_bearing$life_span)
qqnorm(ball_bearing$life_span)

}
\keyword{ball}
\keyword{bearings,}
\keyword{datasets,}
\keyword{inference}
\keyword{means}
\keyword{on}
