% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-package.R
\docType{data}
\name{cherry}
\alias{cherry}
\title{Summary information for 31 cherry trees}
\format{A tibble with 31 observations on the following 3 variables.
\describe{\item{Diam}{diameter in inches (at 54 inches above
ground)} \item{Height}{height is measured in feet}
\item{Volume}{volume in cubic feet} }}
\source{
D.J. Hand. A handbook of small data sets. Chapman & Hall/CRC, 1994.
}
\usage{
cherry
}
\description{
Researchers wanting to understand the relationship between these variables
for black cherry trees collected data from 31 trees in the Allegheny
National Forest, Pennsylvania.
}
\examples{

model <- lm(Volume ~ Diam + Height, cherry)
summary(model)
plot(model)

}
\keyword{datasets}
