% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-package.R
\docType{data}
\name{cle_sac}
\alias{cle_sac}
\alias{cle.sac}
\title{Cleveland and Sacramento}
\format{A tibble with 500 observations representing people on the
following 8 variables.  \describe{\item{year}{Year the data was
collected.} \item{state}{State where person resides.}
\item{city}{City.} \item{age}{Age of the person.}
\item{sex}{Gender.} \item{race}{Ethnicity.}
\item{marital_status}{Marital status.}
\item{personal_income}{Personal income.} }}
\usage{
cle_sac
}
\description{
Data on a sample of 500 people from the Cleveland, OH and Sacramento, CA
metro areas.
}
\examples{

d <- cle_sac[!is.na(cle_sac$personal_income),]
set.seed(8957)
is.sac <- d$city == "Sacramento_ CA"
is.cle <- d$city == "Cleveland_ OH"
sac <- sample(d$personal_income[is.sac], 17)
cle <- sample(d$personal_income[is.cle], 21)
par(mfrow = c(2, 1), las = 1)
histPlot(sac, breaks = 8)
histPlot(cle, breaks = 8)

}
\keyword{datasets}
