% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-package.R
\docType{data}
\name{countycomplete}
\alias{countycomplete}
\title{United States Counties}
\format{A tibble with 3143 observations on the following 53 variables.
\describe{\item{state}{State.} \item{name}{County name.}
\item{FIPS}{FIPS code.} \item{pop2010}{2010 county
population.} \item{pop2000}{2000 county population.}
\item{age_under_5}{Percent of population under 5 (2010).}
\item{age_under_18}{Percent of population under 18 (2010).}
\item{age_over_65}{Percent of population over 65 (2010).}
\item{female}{Percent of population that is female (2010).}
\item{white}{Percent of population that is white (2010).}
\item{black}{Percent of population that is black (2010).}
\item{native}{Percent of population that is a Native American
(2010).} \item{asian}{Percent of population that is a Asian (2010).}
\item{pac_isl}{Percent of population that is Hawaii or Pacific
Islander (2010).} \item{two_plus_races}{Percent of population that
identifies as two or more races (2010).} \item{hispanic}{Percent of
population that is Hispanic (2010).}
\item{white_not_hispanic}{Percent of population that is white and
not Hispanic (2010).} \item{no_move_in_one_plus_year}{Percent of
population that has not moved in at least one year (2006-2010).}
\item{foreign_born}{Percent of population that is foreign-born
(2006-2010).} \item{foreign_spoken_at_home}{Percent of population
that speaks a foreign language at home (2006-2010).}
\item{hs_grad}{Percent of population that is a high school graduate
(2006-2010).} \item{bachelors}{Percent of population that earned a
bachelor's degree (2006-2010).} \item{veterans}{Percent of
population that are veterans (2006-2010).}
\item{mean_work_travel}{Mean travel time to work (2006-2010).}
\item{housing_units}{Number of housing units (2010).}
\item{home_ownership}{Homeownership rate (2006-2010).}
\item{housing_multi_unit}{Housing units in multi-unit structures
(2006-2010).} \item{median_val_owner_occupied}{Median value of
owner-occupied housing units (2006-2010).}
\item{households}{Households (2006-2010).}
\item{persons_per_household}{Persons per household (2006-2010).}
\item{per_capita_income}{Per capita money income in past 12 months
(2010 dollars, 2006-2010)} \item{median_household_income}{Median
household income (2006-2010).} \item{poverty}{Percent below poverty
level (2006-2010).} \item{private_nonfarm_establishments}{Private
nonfarm establishments (2009).}
\item{private_nonfarm_employment}{Private nonfarm employment
(2009).} \item{percent_change_private_nonfarm_employment}{Private
nonfarm employment, percent change from 2000 to 2009.}
\item{nonemployment_establishments}{Nonemployer establishments
(2009).} \item{firms}{Total number of firms (2007).}
\item{black_owned_firms}{Black-owned firms, percent (2007).}
\item{native_owned_firms}{Native American-owned firms, percent
(2007).} \item{asian_owned_firms}{Asian-owned firms, percent
(2007).} \item{pac_isl_owned_firms}{Native Hawaiian and other
Pacific Islander-owned firms, percent (2007).}
\item{hispanic_owned_firms}{Hispanic-owned firms, percent (2007).}
\item{women_owned_firms}{Women-owned firms, percent (2007).}
\item{manufacturer_shipments_2007}{Manufacturer shipments, 2007
($1000).} \item{mercent_whole_sales_2007}{Merchange wholesaler
sales, 2007 ($1000).} \item{sales}{Retail sales, 2007 ($1000).}
\item{sales_per_capita}{Retail sales per capita, 2007.}
\item{accommodation_food_service}{Accommodation and food services
sales, 2007 ($1000).} \item{building_permits}{Building permits
(2010).} \item{fed_spending}{Federal spending (2009).}
\item{area}{Land area in square miles (2010).}
\item{density}{Persons per square mile (2010).} }}
\source{
http://quickfacts.census.gov/qfd/states/ (no longer available)
}
\usage{
countycomplete
}
\description{
Data for 3143 counties in the United States.
}
\examples{

countycomplete


}
\references{
OpenIntro Statistics,
\href{http://www.openintro.org/os}{openintro.org/os}
}
\keyword{datasets}
