% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-package.R
\docType{data}
\name{credits}
\alias{credits}
\title{College credits.}
\format{A tibble with 100 observations on the following variable.
\describe{\item{credits}{Number of credits.} }}
\source{
Simulated data.
}
\usage{
credits
}
\description{
A simulated data set of number of credits taken by college students each
semester.
}
\examples{

histPlot(credits$credits)

}
\keyword{college}
\keyword{credits,}
\keyword{datasets,}
\keyword{inference}
\keyword{means}
\keyword{on}
