% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-marathon.R
\docType{data}
\name{marathon}
\alias{marathon}
\title{New York City Marathon Times}
\format{A tibble with 60 observations on the following 3 variables.
\describe{
\item{year}{Year}
\item{gender}{Gender}
\item{time}{Running time (in hours)}
}}
\source{
\url{http://www.webcitation.org/5kx7ilFLp}
}
\usage{
marathon
}
\description{
Marathon times of male and female winners of the New York City Marathon
1970-1999.
}
\examples{

str(marathon)
histPlot(marathon$time)
boxPlot(marathon$time, horiz = TRUE, fact = marathon$gender)

}
\keyword{datasets}
