% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-package.R
\docType{data}
\name{mlbbat10}
\alias{mlbbat10}
\title{Major League Baseball Player Hitting Statistics for 2010}
\format{A tibble with 1199 observations on the following 19 variables.
\describe{\item{name}{Player name} \item{team}{Team
abbreviation} \item{position}{Player position}
\item{G}{Number of games} \item{AB}{Number of at bats}
\item{R}{Number of runs} \item{H}{Number of hits}
\item{2B}{Number of doubles} \item{3B}{Number of triples}
\item{HR}{Number of home runs} \item{RBI}{Number of runs
batted in} \item{TB}{Total bases, computed as 3\emph{HR + 2}3B + 1*2B +
H} \item{BB}{Number of walks} \item{SO}{Number of
strikeouts} \item{SB}{Number of stolen bases}
\item{CS}{Number of times caught stealing} \item{OBP}{On
base percentage} \item{SLG}{Slugging percentage (TB / AB)}
\item{AVG}{Batting average} }}
\source{
Data was collected from mlb.com on April 22nd, 2011.
}
\usage{
mlbbat10
}
\description{
Major League Baseball Player Hitting Statistics for 2010.
}
\examples{

d   <- mlbbat10[mlbbat10$AB > 200,]
pos <- list(c("OF"), c("1B", "2B", "3B", "SS"), "DH", "C")
POS <- c("OF", "IF", "DH", "C")

#=====> On-base Percentage Across Positions <=====#
out <- c()
gp  <- c()
for(i in 1:length(pos)){
	these <- which(d$pos \%in\% pos[[i]])
	out   <- c(out, d[these,"OBP"])
	gp    <- c(gp, rep(POS[i], length(these)))
}
plot(out ~ as.factor(gp))
summary(lm(out ~ as.factor(gp)))
anova(lm(out ~ as.factor(gp)))

#=====> Batting Average Across Positions <=====#
out <- c()
gp  <- c()
for(i in 1:length(pos)){
	these <- which(d$pos \%in\% pos[[i]])
	out   <- c(out, d[these,"AVG"])
	gp    <- c(gp, rep(POS[i], length(these)))
}
plot(out ~ as.factor(gp))
summary(lm(out ~ as.factor(gp)))
anova(lm(out ~ as.factor(gp)))

#=====> Home Runs Across Positions <=====#
out <- c()
gp  <- c()
for(i in 1:length(pos)){
	these <- which(d$pos \%in\% pos[[i]])
	out   <- c(out, d[these,"HR"])
	gp    <- c(gp, rep(POS[i], length(these)))
}
plot(out ~ as.factor(gp))
summary(lm(out ~ as.factor(gp)))
anova(lm(out ~ as.factor(gp)))

#=====> Runs Batted In Across Positions <=====#
out <- c()
gp  <- c()
for(i in 1:length(pos)){
	these <- which(d$pos \%in\% pos[[i]])
	out   <- c(out, d[these,"RBI"])
	gp    <- c(gp, rep(POS[i], length(these)))
}
plot(out ~ as.factor(gp))
summary(lm(out ~ as.factor(gp)))
anova(lm(out ~ as.factor(gp)))


}
\keyword{datasets}
