% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-package.R
\docType{data}
\name{mtl}
\alias{mtl}
\title{Medial temporal lobe (MTL) and other data for 26 participants}
\format{A tibble with 35 observations on the following 23 variables.
\describe{\item{Subject}{ID for the individual.}
\item{Sex}{Gender, which takes values \code{F} (female) or \code{M}
(male).} \item{Ethnic}{Ethnicity, simplified to \code{Caucasian} and
\code{Other}.} \item{Educ}{Years of educational.}
\item{e4grp}{APOE-4 status, taking a value of \code{E4} or
\code{Non-E4}.} \item{Age}{Age, in years.} \item{MMSE}{
Score from the Mini-Mental State Examination, which is a global cognitition
evaluation.} \item{HamD}{Score on the Hamilton Rating Scale for
depression.} \item{HamA}{Score on the Hamilton Rating Scale for
anxiety.} \item{DigSym}{We (the authors of this R package) are
unsure as to the meaning of this variable.} \item{DelayVP}{We
(the authors of this R package) are unsure as to the meaning of this
variable.} \item{BFRSelective.Reminding.Delayed}{We (the authors
of this R package) are unsure as to the meaning of this variable.}
\item{Sitting}{Self-reported time sitting per day, averaged to the
nearest hour.} \item{METminwk}{Metabolic equivalent units score
(activity level).  A score of \code{0} means \dQuote{no activity} while
\code{3000} is considered \dQuote{high activity}.} \item{IPAQgrp}{
Classification of \code{METminwk} into \code{Low} or \code{High}.}
\item{ACA1}{Thickness of the CA1 subregion of the MTL.}
\item{ACA23DG}{Thickness of the CA23DG subregion of the MTL.}
\item{AECort}{Thickness of a subregion of the MTL.}
\item{AFusiCort}{Thickness of the fusiform gyrus subregion of the
MTL.} \item{APhCort}{Thickness of the perirhinal cortex subregion of
the MTL.} \item{APeCort}{Thickness of the entorhinal cortex
subregion of the MTL.} \item{Asubic}{Thickness of the subiculum
subregion of the MTL.} \item{TOTAL}{Total MTL thickness.} }}
\source{
Siddarth P, Burggren AC, Eyre HA, Small GW, Merrill DA. 2018.
Sedentary behavior associated with reduced medial temporal lobe thickness in
middle-aged and older adults. PLoS ONE 13(4): e0195549.
\url{https://doi.org/10.1371/journal.pone.0195549}

Thank you to Professor Silas Bergen of Winona State University for pointing
us to this data set!
}
\usage{
mtl
}
\description{
The data are from a convenience sample of 25 women and 10 men who were
middle-aged or older. The purpose of the study was to understand the
relationship between sedentary behavior and thickness of the medial temporal
lobe (MTL) in the brain.
}
\examples{


# Examine the relationship between the METminwk and IPAQgrp variables.
a <- mtl[, c("METminwk", "IPAQgrp")]
a[order(a$METminwk), ]


}
\references{
A New York Times article references this study.
\url{https://www.nytimes.com/2018/04/19/opinion/standing-up-at-your-desk-could-make-you-smarter.html}
}
\keyword{datasets}
