% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-package.R
\docType{data}
\name{nba_heights}
\alias{nba_heights}
\title{NBA Player heights from 2008-9}
\format{A tibble with 435 observations (players) on the following 4
variables.  \describe{\item{last.name}{Last name.}
\item{first.name}{First name.} \item{h.meters}{Height, in
meters.} \item{h.in}{Height, in inches.} }}
\source{
Collected from \url{http://www.nba.com}.
}
\usage{
nba_heights
}
\description{
Heights of all NBA players from the 2008-9 season.
}
\examples{

qqnorm(nba_heights$h.meters)

}
\keyword{datasets}
