% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openintro.R
\docType{package}
\name{openintro-package}
\alias{openintro}
\alias{openintro-package}
\title{openintro: OpenIntro data sets and supplemental functions}
\description{
This package is a supplement to OpenIntro Statistics, which is a free textbook
available at \href{http://www.openintro.org/}{openintro.org} (at-cost
paperbacks are also available for under $10 on Amazon). The package contains
data sets used in the textbook along with custom plotting functions for
reproducing book figures. Note that many functions and examples include color
transparency. Some plotting elements may not show up properly (or at all) in
some Windows versions.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.openintro.org/}
  \item Report bugs at \url{https://github.com/OpenIntroStat/openintro-r-package/issues}
}

}
\author{
\strong{Maintainer}: Mine Çetinkaya-Rundel \email{mine@rstudio.com} (0000-0001-6452-2420)

Authors:
\itemize{
  \item David Diez \email{david.m.diez@gmail.com}
  \item Andrew Bray \email{abray@reed.edu}
  \item Christopher Barr
}

}
