% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-package.R
\docType{data}
\name{outliers}
\alias{outliers}
\title{Simulated data sets for different types of outliers}
\format{A tibble with 50 observations on the following 5 variables.
\describe{\item{x}{a numeric vector} \item{y}{a numeric
vector} \item{xInf}{a numeric vector} \item{yLev}{a numeric
vector} \item{yOut}{a numeric vector} }}
\usage{
outliers
}
\description{
Data sets for showing different types of outliers
}
\examples{

outliers

}
\keyword{datasets}
