% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-package.R
\docType{data}
\name{penny_ages}
\alias{penny_ages}
\alias{penny.ages}
\title{Penny Ages}
\format{A tibble with 648 observations on the following 2 variables.
\describe{\item{Year}{Penny's year.} \item{age}{Age as of
2004.} }}
\usage{
penny_ages
}
\description{
Sample of pennies and their ages. Taken in 2004.
}
\examples{

hist(penny_ages$Year)

}
\keyword{datasets}
