% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-package.R
\docType{data}
\name{possum}
\alias{possum}
\title{possum}
\format{A tibble with 104 observations on the following 8 variables.
\describe{\item{site}{The site number where the possum was
trapped. } \item{pop}{Population, either \code{Vic} (Victoria) or
\code{other} (New South Wales or Queensland).} \item{sex}{Gender,
either \code{m} (male) or \code{f} (female).} \item{age}{Age. }
\item{headL}{Head length, in mm. } \item{skullW}{Skull
width, in mm. } \item{totalL}{Total length, in cm. }
\item{tailL}{Tail length, in cm. } }}
\source{
Lindenmayer, D. B., Viggers, K. L., Cunningham, R. B., and Donnelly,
C. F. 1995. Morphological variation among columns of the mountain brushtail
possum, Trichosurus caninus Ogilby (Phalangeridae: Marsupiala). Australian
Journal of Zoology 43: 449-458.
}
\usage{
possum
}
\description{
Data representing possums in Australia and New Guinea. This is a copy of the
data set by the same name in the \code{DAAG} package, however, the data set
included here includes fewer variables.
}
\examples{

plot(possum$headL, possum$skullW)
densityPlot(possum$totalL, possum$sex, key=c('f','m'),
	xlab='total length (cm)')
legend('topright', col=c('black', 'red'), lty=1:2, legend=c('f', 'm'))

}
\references{
\url{http://www.openintro.org/}
}
\keyword{datasets}
