% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-package.R
\docType{data}
\name{"speed_gender_height"}
\alias{"speed_gender_height"}
\alias{speed_gender_height}
\title{Speed, gender, and height of 1302 students}
\format{A tibble with 1325 observations on the following 4 variables.
\describe{\item{X}{a numeric vector} \item{speed}{a numeric
vector} \item{gender}{a factor with levels \code{female}
\code{male}} \item{height}{a numeric vector} }}
\usage{
speed_gender_height
}
\description{
1,302 UCLA students were asked to fill out a survey where they were asked
about their height, fastest speed they have ever driven, and gender.
}
\examples{

speed_gender_height

}
\keyword{datasets}
