% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-package.R
\docType{data}
\name{satgpa}
\alias{satgpa}
\title{SAT and GPA data}
\format{A tibble with 1000 observations on the following 6 variables.
\describe{\item{sex}{Gender of the student.}
\item{SATV}{Verbal SAT percentile.} \item{SATM}{Math SAT
percentile.} \item{SATSum}{Total of verbal and math SAT
percentiles.} \item{HSGPA}{High school grade point average.}
\item{FYGPA}{First year (college) grade point average.} }}
\source{
Educational Testing Service originally collected the data.
}
\usage{
satgpa
}
\description{
SAT and GPA data for 1000 students at an unnamed college.
}
\examples{

par(mfrow=2:1)

plot(satgpa$SATSum/2, satgpa$FYGPA)
g <- lm(satgpa$FYGPA ~ I(satgpa$SATSum/2))
summary(g)
abline(g)

plot(satgpa$SATM, satgpa$FYGPA)
g <- lm(satgpa$FYGPA ~ satgpa$SATM)
summary(g)
abline(g)

}
\references{
Data retrieved from

\url{https://www.dartmouth.edu/~chance/course/Syllabi/Princeton96/Class12.html}

Data utilized in Chapter 7 of the Open Intro Statistics book:
\url{http://www.openintro.org/}
}
\keyword{datasets}
