% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-package.R
\docType{data}
\name{tips}
\alias{tips}
\title{Tip data}
\format{A tibble with 95 observations on the following 5 variables.
\describe{\item{week}{Week number. } \item{day}{Day,
either \code{Friday} or \code{Tuesday}. } \item{nPeop}{Number of
people associated with the group. } \item{bill}{Total bill for the
group. } \item{tip}{Total tip from the group. } }}
\source{
Simulated data set.
}
\usage{
tips
}
\description{
A simulated data set of tips over a few weeks on a couple days per week.
Each tip is associated with a single group, which may include several bills
and tables (i.e. groups paid in one lump sum in simulations).
}
\details{
This data set was built using simulations of tables, then bills, then tips
based on the bills. Large groups were assumed to only pay the gratuity,
which is evident in the data. Tips were set to be plausible round values;
they were often (but not always) rounded to dollars, quarters, etc.
}
\examples{

par(mfrow=c(2,2))
boxPlot(tips$tip, tips$day)
densityPlot(tips$tip, tips$week, key=1:3)
legend('topright', lty=1:3, col=c('black', 'red', 'blue'), legend=1:3)
dotPlot(tips$tip)
densityPlot(tips$tip, tips$day)
legend('topright', col=c('black','red'), lty=1:2,
	legend=c('Tuesday', 'Friday'))

}
\references{
\url{http://www.openintro.org/}
}
\keyword{datasets}
