% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-package.R
\docType{data}
\name{trees}
\alias{trees}
\title{Summary of 32 trees}
\format{A tibble with 31 observations on the following 3 variables.
\describe{\item{Girth}{Diameter, in inches}
\item{Height}{Height, in feet.} \item{Volume}{Volume, in
cubic feet.} }}
\source{
Source: R Dataset.
http://stat.ethz.ch/R-manual/R-patched/library/datasets/html/trees.html
}
\usage{
trees
}
\description{
Variables include girth, height, and volume.
}
\examples{

trees

}
\keyword{datasets}
