% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-package.R
\docType{data}
\name{urban_rural_pop}
\alias{urban_rural_pop}
\title{State summary info}
\format{A tibble with 51 observations on the following 5 variables.
\describe{\item{State}{US state.} \item{UrbanIn}{a numeric
vector} \item{UrbanOut}{a numeric vector} \item{RuralFarm}{a
numeric vector} \item{RuralNonfarm}{a numeric vector} }}
\source{
US census.
}
\usage{
urban_rural_pop
}
\description{
Census info for the 50 US states plus DC.
}
\examples{

urban_rural_pop

}
\keyword{datasets}
