% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-package.R
\docType{data}
\name{xom}
\alias{xom}
\title{Exxon Mobile stock data}
\format{A tibble with 98 observations on the following 7 variables.
\describe{\item{Date}{Date.} \item{Open}{a numeric vector}
\item{High}{a numeric vector} \item{Low}{a numeric vector}
\item{Close}{a numeric vector} \item{Volume}{a numeric
vector} \item{Adj.Close}{a numeric vector} }}
\source{
http://finance.yahoo.com
}
\usage{
xom
}
\description{
Monthly data covering 2006 through early 2014.
}
\examples{

xom

}
\keyword{datasets}
