% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-package.R
\docType{data}
\name{census}
\alias{census}
\title{Random sample of 2000 U.S. Census Data}
\format{A tibble with 500 observations on the following 8 variables.
\describe{
\item{censusYear}{Census Year.}
\item{stateFIPScode}{Name of state.}
\item{totalFamilyIncome}{Total family income (in U.S. dollars).}
\item{age}{Age.}
\item{sex}{Sex with levels \code{Female} and \code{Male}.}
\item{raceGeneral}{Race with levels \code{American
Indian or Alaska Native}, \code{Black}, \code{Chinese}, \code{Japanese},
\code{Other Asian or Pacific Islander}, \code{Two major races}, \code{White}
and \code{Other}.}
\item{maritalStatus}{Marital status with levels
\code{Divorced}, \code{Married/spouse absent}, \code{Married/spouse
present}, \code{Never married/single}, \code{Separated} and \code{Widowed}.}
\item{totalPersonalIncome}{Total personal income (in U.S. dollars).}
}}
\source{
\url{http://factfinder.census.gov/}
}
\usage{
census
}
\description{
A random sample of 500 observations from the 2000 U.S. Census Data.
}
\examples{

str(census)
these <- census[,3] > 0     # income greater than 0
histPlot(log(census$totalFamilyIncome[these]), xlab="log(total family income)")

}
\keyword{datasets}
