% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ipod.R
\docType{data}
\name{ipod}
\alias{ipod}
\title{Length of songs on an iPod}
\format{A tibble with 3000 observations on the following variable.
\describe{
\item{song_length}{Length of song (in minutes).}
}}
\source{
Simulated data.
}
\usage{
ipod
}
\description{
A simulated data set on lengths of songs on an iPod.
}
\examples{

histPlot(ipod$song_length)

}
\keyword{datasets,}
\keyword{iPod,}
\keyword{inference}
\keyword{means}
\keyword{on}
